refactor: move ecstore rebalance type contracts
